requirements.txt
bottle
canister
pyjwt
ed25519
base58
requests
uuid
sentry_sdk